requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.3
blinker==1.9.0
cachetools==6.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comtypes==1.4.10
cryptography==45.0.4
distro==1.9.0
ecdsa==0.19.1
elevenlabs==1.57.0
et_xmlfile==2.0.0
Flask==3.1.0
flask-cors==6.0.0
gitdb==4.0.12
GitPython==3.1.44
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
MarkupSafe==3.0.2
msal==1.32.3
narwhals==1.43.1
numpy==2.2.2
openai==1.76.2
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
pillow==11.2.1
protobuf==6.31.1
pyarrow==20.0.0
PyAudio==0.2.14
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
pydeck==0.9.1
PyJWT==2.10.1
pypiwin32==223
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-http-client==3.3.7
pyttsx3==2.98
pytz==2025.1
pywin32==310
referencing==0.36.2
requests==2.32.3
rpds-py==0.25.1
sendgrid==6.12.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
SpeechRecognition==3.14.2
streamlit==1.46.0
tenacity==9.1.2
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
watchdog==6.0.0
websockets==15.0.1
Werkzeug==3.1.3
